+++ rust/aura-arch/src/lib.rs
@@ -8,6 +8,9 @@ pub const DEFAULT_ROOT: &str = "/";
 /// The default filepath of the pacman/libalpm database.
 pub const DEFAULT_DB: &str = "/var/lib/pacman/";
 
+/// The default filepath of the pacman configuration.
+pub const DEFAULT_PAC_CONF: &str = "/etc/pacman.conf";
+
 #[derive(Debug)]
 pub enum Error {
     Alpm(alpm::Error),