+++ rust/aura/src/command.rs
@@ -1,3 +1,4 @@
 //! All Aura commands.
 
+pub mod misc;
 pub mod orphans;